pyproject.toml
[project]
name = "fastapi-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.1",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "rank-bm25>=0.2.2",
]
